+++ Dfe.Academies.Application/Establishment/IEstablishmentQueries.cs
@@ -1,6 +1,6 @@
 ﻿using Dfe.Academies.Contracts.V4.Establishments;
 
-namespace Dfe.Academies.Application.Queries.Establishment
+namespace Dfe.Academies.Application.Establishment
 {
     public interface IEstablishmentQueries
     {
@@ -8,7 +8,7 @@ namespace Dfe.Academies.Application.Queries.Establishment
         Task<EstablishmentDto?> GetByUrn(string urn, CancellationToken cancellationToken);
         Task<(List<EstablishmentDto>, int)> Search(string name, string ukPrn, string urn, CancellationToken cancellationToken);
         Task<IEnumerable<int>> GetURNsByRegion(string[] regions, CancellationToken cancellationToken);
-        Task<List<EstablishmentDto>> GetByUrns(int[] Urns, CancellationToken cancellationToken);       
+        Task<List<EstablishmentDto>> GetByUrns(int[] Urns, CancellationToken cancellationToken);
         Task<List<EstablishmentDto>> GetByTrust(string trustUkprn, CancellationToken cancellationToken);
     }
 }
